Reset path_interpolator, clearpath error event to improve error handling

+++ software/delta-control/Src/drivers/configuration.c
@@ -377,7 +377,7 @@ config_set_cpu_load( uint8_t percent )
 }
 
 PUBLIC void
-config_set_cpu_clock( float clock )
+config_set_cpu_clock(uint32_t clock)
 {
 	sys_stats.cpu_clock = clock / 1000000;	//convert to Mhz
 }
